update global image repo config
